MNT #31 diffractometer.export
